fork off build notifier daemon directly
Instead of spawning a child process that daemonizes itself.
This reduces overhead and allows us to share the FIFO fd
and paths to lock directly instead of on the commmand line.

Make sure we do not allocate in the daemon because
it has no stderr to log errors to, complicating debugging.

Remove debug logging before emitting IPC messages into the FIFO
because that was not very helpful and forced us to allocate.